Fix uncertainty in the load of streams
Currently we don't know in which space the streamlines are after being
loaded. In this commit, we add an option to the load function to state
if the streamlines should be in voxel or mm space. If the voxel space is
selected, then we check if the affine is correctly set.